refactor(task): decompose cycle resolution helpers and extract graph utilities into graphresolver_graphutil.go
